+++ bin/docker-entrypoint.sh
@@ -2,4 +2,4 @@
 
 set -e
 
-cd /controls && npm install && node server
+cd /controls && npm install && grunt && node server