+++ js/arrange.js
@@ -21,6 +21,7 @@
         defaults = {
           itemsSelector: null,
           connectWith: false,
+          placeholder: null,
           placeholderCssClass: 'arrange-placeholder'
         },
         settings = $.extend({}, defaults, options);
@@ -57,6 +58,10 @@
           placeholder = $('<'+ items.first()[0].tagName +' />');
         }
 
+        if (settings.placeholder) {
+          placeholder = $(settings.placeholder);
+        }
+
         self.dragStart = 'dragstart.arrange touchstart.arrange gesturestart.arrange';
         self.dragEnd = 'dragend.arrange touchend.arrange touchcancel.arrange gestureend.arrange';
         self.dragWhileDragging = 'dragover.arrange dragenter.arrange drop.arrange touchmove.arrange gesturechange.arrange';
@@ -77,8 +82,10 @@
             .data('connectWith', self.connectWith);
         }
 
+        self.items = items;
+
         // Draggable Items
-        items
+        self.items
         .attr('draggable', true).addClass(self.handle ? '' : 'draggable')
         .add([this, placeholder])
         .not('a[href], img').on('selectstart.arrange', function() {
@@ -171,7 +178,7 @@
             }
             else if (!self.placeholders.is(this)) {
               self.placeholders.detach();
-              this.element.append(placeholder);
+              self.element.append(placeholder);
             }
             return false;
           });//-------------------------------------------------------------------------------------
@@ -191,13 +198,28 @@
         return returns;
       },
 
+      unbind: function() {
+        this.items
+          .removeClass('draggable')
+          .removeAttr('draggable')
+          .off('selectstart.arrange '+ this.dragStart +' '+ this.dragEnd +' '+ this.dragWhileDragging);
+
+        $(this.handle, this.items)
+          .removeClass('draggable')
+          .off('mousedown.arrange mouseup.arrange touchstart.arrange touchend.arrange');
+
+        return this;
+      },
+
+      updated: function() {
+        return this
+          .unbind()
+          .init();
+      },
+
       // Teardown
       destroy: function() {
-        var items = (this.connectWith) ?
-          this.element.children().add($(this.connectWith).children()) : this.element.children();
-
-        items.off('selectstart.arrange '+ this.dragStart +' '+ this.dragEnd +' '+ this.dragWhileDragging);
-        $(this.handle, items).off('mousedown.arrange mouseup.arrange touchstart.arrange touchend.arrange');
+        this.unbind();
         $.removeData(this.element[0], pluginName);
       }
     };
@@ -206,8 +228,8 @@
     return this.each(function() {
       var instance = $.data(this, pluginName);
       if (instance) {
-        instance.settings = $.extend({}, defaults, options);
-        instance.show();
+        instance.settings = $.extend({}, instance.settings, options);
+        instance.updated();
       } else {
         instance = $.data(this, pluginName, new Plugin(this, settings));
       }