+++ js/datagrid.js
@@ -1553,11 +1553,7 @@ $.fn.datagrid = function(options) {
       }
 
       //Attach Keyboard support
-      var popupOpts = {
-        offset: {
-          y: 15
-        }
-      };
+      var popupOpts = {offset: {y: 15}, placementOpts: {strategies: ['flip', 'nudge']}};
 
       this.headerRow.addClass('is-filterable');
       this.headerRow.find('.btn-filter').popupmenu(popupOpts).on('selected.datagrid', function () {
@@ -3069,6 +3065,20 @@ $.fn.datagrid = function(options) {
       this.headerWidths = [];
     },
 
+    //size the first row
+    resyncColumnWidths: function () {
+      if (!this.hasFixedHeader) {
+        return;
+      }
+
+      for (var j = 0; j < this.settings.columns.length; j++) {
+        var column = this.settings.columns[j],
+          node = this.tableBody.find('tr:visible:first td').eq(j);
+
+        node.attr('style', this.calculateColumnWidth(column, true, j).replace('style="', '').replace('"', ''));
+      }
+    },
+
     // Get child offset
     getChildOffset: function(obj) {
       var childPos = obj.offset(),
@@ -3263,7 +3273,7 @@ $.fn.datagrid = function(options) {
           cell = elem.parent().children(':visible').index(elem),
           rowNode = $(this).closest('tr'),
           dataRowIdx = self.dataRowIndex(rowNode),
-          col = self.columnSettings(cell),
+          col = self.columnSettings(cell, true),
           item = self.settings.treeGrid ?
             self.settings.treeDepth[dataRowIdx].node :
             self.settings.dataset[self.pager && self.settings.source ? rowNode.index() : dataRowIdx];
@@ -3577,6 +3587,10 @@ $.fn.datagrid = function(options) {
           menu.append('<li><a href="#" data-option="personalize-columns">' + Locale.translate('PersonalizeColumns') + '</a></li>');
         }
 
+        if (settings.toolbar.resetLayout) {
+          menu.append('<li><a href="#" data-option="reset-layout">' + Locale.translate('ResetDefault') + '</a></li>');
+        }
+
         if (settings.toolbar.exportToExcel) {
           menu.append('<li><a href="#" data-option="export-to-excel">' + Locale.translate('ExportToExcel') + '</a></li>');
         }
@@ -3623,6 +3637,10 @@ $.fn.datagrid = function(options) {
 
         if (action === 'personalize-columns') {
           self.personalizeColumns();
+        }
+
+        if (action === 'reset-layout') {
+          self.resetColumns();
         }
 
         if (action === 'export-to-excel') {
@@ -3741,19 +3759,20 @@ $.fn.datagrid = function(options) {
 
           });
 
-        // Hide non matching rows
-
-        if (!found) {
-          row.addClass('is-filtered').hide();
-        } else if (found && row.is('.datagrid-expandable-row')) {
-          row.prev().show();
-          row.prev().find('.datagrid-expand-btn').addClass('is-expanded');
-          row.prev().find('.plus-minus').addClass('active');
-          row.addClass('is-expanded').css('display', 'table-row');
-          row.find('.datagrid-row-detail').css('height', 'auto');
-        }
+          // Hide non matching rows
+          if (!found) {
+            row.addClass('is-filtered').hide();
+          } else if (found && row.is('.datagrid-expandable-row')) {
+            row.prev().show();
+            row.prev().find('.datagrid-expand-btn').addClass('is-expanded');
+            row.prev().find('.plus-minus').addClass('active');
+            row.addClass('is-expanded').css('display', 'table-row');
+            row.find('.datagrid-row-detail').css('height', 'auto');
+          }
 
       });
+
+      this.resyncColumnWidths();
     },
 
     //Get or Set Selected Rows
@@ -4715,8 +4734,12 @@ $.fn.datagrid = function(options) {
     //Returns Column Settings from a cell
     firstRow: null,
 
-    columnSettings: function (cell) {
+    columnSettings: function (cell, onlyVisible) {
       var column = settings.columns[cell];
+
+      if (onlyVisible) {
+        column = this.visibleColumns()[cell];
+      }
 
       return column || {};
     },