+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Release Date: 2017-01-04
 ### Affects
 - Datagrid (Columns)
 - Toolbar Searchfield
+- Toolbar (scroll bar)
+- Auto Complete / Editor (xss)
 
 Jira Release Notes
 http://jira.infor.com/secure/ReleaseNote.jspa?version=27962&styleName=Html&projectId=10980&Create=Create&atl_token=ATP9-LKKS-XFKU-5RYX%7Cdfaec3f836ea373670a7da5f351101b3cdec20fd%7Clin